kvutils/tools: Move `WriteSetEntries` logic into CommitStrategySupport.
This makes it more flexible with regards to extended write sets.
