+++ ledger/participant-state/kvutils/tools/integrity-check/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/LogAppendingCommitStrategySupport.scala
@@ -6,12 +6,20 @@ package com.daml.ledger.participant.state.kvutils.tools.integritycheck
 import com.codahale.metrics.MetricRegistry
 import com.daml.ledger.on.memory.{InMemoryLedgerStateOperations, Index}
 import com.daml.ledger.participant.state.kvutils
+import com.daml.ledger.participant.state.kvutils.DamlKvutils.{
+  DamlLogEntry,
+  DamlLogEntryId,
+  DamlStateKey,
+  DamlStateValue
+}
+import com.daml.ledger.participant.state.kvutils.Envelope
 import com.daml.ledger.participant.state.kvutils.tools.integritycheck.IntegrityChecker.bytesAsHexString
 import com.daml.ledger.validator.LedgerStateOperations.{Key, Value}
 import com.daml.ledger.validator.batch.BatchedSubmissionValidatorFactory
 import com.daml.ledger.validator.{
   CommitStrategy,
   DamlLedgerStateReader,
+  DefaultStateKeySerializationStrategy,
   StateKeySerializationStrategy
 }
 import com.daml.metrics.Metrics
@@ -60,4 +68,28 @@ final class LogAppendingCommitStrategySupport(implicit executionContext: Executi
         s"Expected: $expectedLogEntry${System.lineSeparator()}Actual: $actualLogEntry"
     )
   }
+
+  override def checkEntryIsReadable(keyBytes: Key, valueBytes: Value): Either[String, Unit] =
+    Envelope.open(valueBytes) match {
+      case Left(errorMessage) =>
+        Left(s"Invalid value envelope: $errorMessage")
+      case Right(Envelope.LogEntryMessage(logEntry)) =>
+        val _ = DamlLogEntryId.parseFrom(keyBytes)
+        if (logEntry.getPayloadCase == DamlLogEntry.PayloadCase.PAYLOAD_NOT_SET)
+          Left("Log entry payload not set.")
+        else
+          Right(())
+      case Right(Envelope.StateValueMessage(value)) =>
+        val key = DefaultStateKeySerializationStrategy.deserializeStateKey(keyBytes)
+        if (key.getKeyCase == DamlStateKey.KeyCase.KEY_NOT_SET)
+          Left("State key not set.")
+        else if (value.getValueCase == DamlStateValue.ValueCase.VALUE_NOT_SET)
+          Left("State value not set.")
+        else
+          Right(())
+      case Right(Envelope.SubmissionMessage(submission)) =>
+        Left(s"Unexpected submission message: $submission")
+      case Right(Envelope.SubmissionBatchMessage(batch)) =>
+        Left(s"Unexpected submission batch message: $batch")
+    }
 }
