+++ ledger/participant-state/kvutils/tools/src/main/scala/ledger/participant/state/kvutils/tools/integritycheck/IntegrityChecker.scala
@@ -26,7 +26,6 @@ import com.daml.ledger.validator.batch.{
   BatchedSubmissionValidatorParameters,
   ConflictDetection
 }
-import com.daml.ledger.validator.{CommitStrategy, DamlLedgerStateReader}
 import com.daml.lf.engine.{Engine, EngineConfig}
 import com.daml.logging.LoggingContext
 import com.daml.logging.LoggingContext.newLoggingContext
@@ -107,9 +106,6 @@ class IntegrityChecker[LogResult](commitStrategySupport: CommitStrategySupport[L
       _ <- processSubmissions(
         importer,
         submissionValidator,
-        commitStrategySupport.ledgerStateReader,
-        commitStrategySupport.commitStrategy,
-        commitStrategySupport.writeSet,
         expectedReadServiceFactory,
         actualReadServiceFactory,
         config,
@@ -196,9 +192,6 @@ class IntegrityChecker[LogResult](commitStrategySupport: CommitStrategySupport[L
   private def processSubmissions(
       importer: LedgerDataImporter,
       submissionValidator: BatchedSubmissionValidator[LogResult],
-      reader: DamlLedgerStateReader,
-      commitStrategy: CommitStrategy[LogResult],
-      queryableWriteSet: QueryableWriteSet,
       expectedReadServiceFactory: ReplayingReadServiceFactory,
       actualReadServiceFactory: ReplayingReadServiceFactory,
       config: Config,
@@ -216,7 +209,7 @@ class IntegrityChecker[LogResult](commitStrategySupport: CommitStrategySupport[L
           )
           expectedWriteSet.foreach {
             case (key, value) =>
-              val result = WriteSetEntries.checkReadable(key, value)
+              val result = commitStrategySupport.checkEntryIsReadable(key, value)
               result.left.foreach { message =>
                 throw new UnreadableWriteSetException(message)
               }
@@ -228,10 +221,10 @@ class IntegrityChecker[LogResult](commitStrategySupport: CommitStrategySupport[L
               submissionInfo.correlationId,
               submissionInfo.recordTimeInstant,
               submissionInfo.participantId,
-              reader,
-              commitStrategy,
+              commitStrategySupport.ledgerStateReader,
+              commitStrategySupport.commitStrategy,
             ) map { _ =>
-              val actualWriteSet = queryableWriteSet.getAndClearRecordedWriteSet()
+              val actualWriteSet = commitStrategySupport.writeSet.getAndClearRecordedWriteSet()
               val orderedActualWriteSet =
                 if (config.sortWriteSet)
                   actualWriteSet.sortBy(_._1.asReadOnlyByteBuffer())
@@ -376,7 +369,8 @@ object IntegrityChecker {
 
   def run[LogResult](
       config: Config,
-      commitStrategySupportFactory: ExecutionContext => CommitStrategySupport[LogResult]): Unit = {
+      commitStrategySupportFactory: ExecutionContext => CommitStrategySupport[LogResult],
+  ): Unit = {
     runAsync(config, commitStrategySupportFactory).failed
       .foreach {
         case exception: CheckFailedException =>
